[FLINK-9846] [table] Refactor KafkaTableSourceFactory to KafkaTableSourceSinkFactory
